+++ sdk/nodejs/deployKey.ts
@@ -99,6 +99,13 @@ export class DeployKey extends pulumi.CustomResource {
             inputs["project"] = args ? args.project : undefined;
             inputs["title"] = args ? args.title : undefined;
         }
+        if (!opts) {
+            opts = {}
+        }
+
+        if (!opts.version) {
+            opts.version = utilities.getVersion();
+        }
         super(DeployKey.__pulumiType, name, inputs, opts);
     }
 }