+++ sdk/nodejs/getProject.ts
@@ -20,8 +20,15 @@ import * as utilities from "./utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-gitlab/blob/master/website/docs/d/project.html.markdown.
  */
-export function getProject(args: GetProjectArgs, opts?: pulumi.InvokeOptions): Promise<GetProjectResult> {
-    return pulumi.runtime.invoke("gitlab:index/getProject:getProject", {
+export function getProject(args: GetProjectArgs, opts?: pulumi.InvokeOptions): Promise<GetProjectResult> & GetProjectResult {
+    if (!opts) {
+        opts = {}
+    }
+
+    if (!opts.version) {
+        opts.version = utilities.getVersion();
+    }
+    const promise: Promise<GetProjectResult> = pulumi.runtime.invoke("gitlab:index/getProject:getProject", {
         "archived": args.archived,
         "defaultBranch": args.defaultBranch,
         "description": args.description,
@@ -39,6 +46,8 @@ export function getProject(args: GetProjectArgs, opts?: pulumi.InvokeOptions): P
         "webUrl": args.webUrl,
         "wikiEnabled": args.wikiEnabled,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**