+++ sdk/nodejs/pipelineSchedule.ts
@@ -116,6 +116,13 @@ export class PipelineSchedule extends pulumi.CustomResource {
             inputs["project"] = args ? args.project : undefined;
             inputs["ref"] = args ? args.ref : undefined;
         }
+        if (!opts) {
+            opts = {}
+        }
+
+        if (!opts.version) {
+            opts.version = utilities.getVersion();
+        }
         super(PipelineSchedule.__pulumiType, name, inputs, opts);
     }
 }