+++ sdk/python/pulumi_gitlab/project.py
@@ -199,6 +199,10 @@ class Project(pulumi.CustomResource):
         __props__['ssh_url_to_repo'] = None
         __props__['web_url'] = None
 
+        if opts is None:
+            opts = pulumi.ResourceOptions()
+        if opts.version is None:
+            opts.version = utilities.get_version()
         super(Project, __self__).__init__(
             'gitlab:index/project:Project',
             resource_name,