+++ sdk/python/pulumi_gitlab/project_membership.py
@@ -60,6 +60,10 @@ class ProjectMembership(pulumi.CustomResource):
             raise TypeError("Missing required property 'user_id'")
         __props__['user_id'] = user_id
 
+        if opts is None:
+            opts = pulumi.ResourceOptions()
+        if opts.version is None:
+            opts.version = utilities.get_version()
         super(ProjectMembership, __self__).__init__(
             'gitlab:index/projectMembership:ProjectMembership',
             resource_name,