+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## HEAD (Unreleased)
-___HEAD___
+* Update to pulumi-terraform@013b95b1c8
 
 ## 0.18.3 (2019-07-09)
 * Fix detailed diffs with nested computed values.